+++ src/App.js
@@ -5,7 +5,7 @@ import Menu from "./components/menu";
 import Header from "./components/header";
 import Sobre from "./components/sobre";
 import Portfolio from "./components/portfolio/";
-import Contato from "./components/contato/contato";
+import Contato from "./components/contato";
 import Footer from "./components/footer";
 
 function App() {
